chore: fixed page refreshes after clone and transfer, added new form--hidden-inputs controller and removed functionality from infinite-scroll controller
