+++ app/components/viral/dialog_component.html.erb
@@ -1,6 +1,7 @@
 <%= render Viral::BaseComponent.new(tag: :span, data: {
   controller: "viral--dialog",
   "viral--dialog-open-value": open,
+  "turbo-permanent": true
 }) do %>
 
   <% if trigger.present? %>
