+++ app/components/viral/flash_component.html.erb
@@ -2,33 +2,68 @@
   data-controller="viral--flash"
   data-viral--flash-timeout-value="<%= timeout %>"
   data-viral--flash-type-value="<%= type %>"
-  class="flex items-center w-full max-w-xs p-4 mb-4 text-slate-500 bg-white rounded-lg shadow dark:text-slate-400 dark:bg-slate-800"
+  class="
+    flex items-center w-full max-w-xs p-4 mb-4 text-slate-500 bg-white rounded-lg
+    shadow dark:text-slate-400 dark:bg-slate-800
+  "
   role="alert"
+  data-turbo-permanent
 >
   <% case type.to_s %>
   <% when "success" %>
-    <div class="inline-flex items-center justify-center flex-shrink-0 w-8 h-8 text-green-500 bg-green-100 rounded-lg dark:bg-green-800 dark:text-green-200">
+    <div
+      class="
+        inline-flex items-center justify-center flex-shrink-0 w-8 h-8 text-green-500
+        bg-green-100 rounded-lg dark:bg-green-800 dark:text-green-200
+      "
+    >
       <%= viral_icon(name: "check_circle_solid", classes: "w-5 h-5") %>
       <span class="sr-only"><%= t("components.flash.success_icon") %></span>
     </div>
   <% when "error" %>
-    <div class="inline-flex items-center justify-center flex-shrink-0 w-8 h-8 text-red-500 bg-red-100 rounded-lg dark:bg-red-800 dark:text-red-200">
+    <div
+      class="
+        inline-flex items-center justify-center flex-shrink-0 w-8 h-8 text-red-500
+        bg-red-100 rounded-lg dark:bg-red-800 dark:text-red-200
+      "
+    >
       <%= viral_icon(name: "x_circle_solid", classes: "w-5 h-5") %>
       <span class="sr-only"><%= t("components.flash.error_icon") %></span>
     </div>
   <% when "warning" %>
-    <div class="inline-flex items-center justify-center flex-shrink-0 w-8 h-8 text-orange-500 bg-orange-100 rounded-lg dark:bg-orange-700 dark:text-orange-200">
+    <div
+      class="
+        inline-flex items-center justify-center flex-shrink-0 w-8 h-8 text-orange-500
+        bg-orange-100 rounded-lg dark:bg-orange-700 dark:text-orange-200
+      "
+    >
       <%= viral_icon(name: "exclamation_circle_solid", classes: "w-5 h-5") %>
       <span class="sr-only"><%= t("components.flash.warning_icon") %></span>
     </div>
   <% when "info" %>
-    <div class="inline-flex items-center justify-center flex-shrink-0 w-8 h-8 text-blue-500 bg-blue-100 rounded-lg dark:bg-blue-800 dark:text-blue-200">
+    <div
+      class="
+        inline-flex items-center justify-center flex-shrink-0 w-8 h-8 text-blue-500
+        bg-blue-100 rounded-lg dark:bg-blue-800 dark:text-blue-200
+      "
+    >
       <%= viral_icon(name: "information_circle_solid", classes: "w-5 h-5") %>
       <span class="sr-only"><%= t("components.flash.information_icon") %></span>
     </div>
   <% end %>
   <div class="ml-3 text-sm font-normal"><%= data %></div>
-  <button type="button" data-action="viral--flash#dismiss" class="ml-auto -mx-1.5 -my-1.5 bg-white text-slate-400 hover:text-slate-900 rounded-lg focus:ring-2 focus:ring-slate-300 p-1.5 hover:bg-slate-100 inline-flex items-center justify-center h-8 w-8 dark:text-slate-500 dark:hover:text-white dark:bg-slate-800 dark:hover:bg-slate-700" data-dismiss-target="#toast-success" aria-label="Close">
+  <button
+    type="button"
+    data-action="viral--flash#dismiss"
+    class="
+      ml-auto -mx-1.5 -my-1.5 bg-white text-slate-400 hover:text-slate-900 rounded-lg
+      focus:ring-2 focus:ring-slate-300 p-1.5 hover:bg-slate-100 inline-flex
+      items-center justify-center h-8 w-8 dark:text-slate-500 dark:hover:text-white
+      dark:bg-slate-800 dark:hover:bg-slate-700
+    "
+    data-dismiss-target="#toast-success"
+    aria-label="Close"
+  >
     <%= viral_icon(name: "x_mark", classes: "w-5 h-5") %>
     <span class="sr-only"><%= t(:"general.screen_reader.close") %></span>
   </button>
