+++ app/views/data_exports/_new_sample_export_dialog.html.erb
@@ -5,7 +5,6 @@
       <div
         data-controller="infinite-scroll"
         data-infinite-scroll-selection-outlet='#samples-table'
-        data-infinite-scroll-field-name-value="data_export[export_parameters][ids][]"
         data-infinite-scroll-paged-field-name-value="sample_ids[]"
         data-infinite-scroll-singular-value="<%= t(".description.singular") %>"
         data-infinite-scroll-plural-value="<%= t(".description.plural") %>"
@@ -46,7 +45,14 @@
             </div>
           </div>
 
-          <%= form_for(:data_export, url: data_exports_path, method: :post) do |form| %>
+          <%= form_for(:data_export, url: data_exports_path, method: :post,
+              data: {
+                controller: "spinner form--hidden-inputs",
+                "form--hidden-inputs-selection-outlet": '#samples-table',
+                "form--hidden-inputs-field-name-value": "data_export[export_parameters][ids][]",
+                action: "turbo:submit-start->spinner#submitStart turbo:submit-end->spinner#submitEnd"
+              }
+            ) do |form| %>
             <div class="grid gap-4">
               <div class="form-field mt-2">
                 <%= form.label :name, t(".name_label") %>
@@ -58,9 +64,9 @@
                                  checked: false,
                                  class:
                                    "w-4 h-4 mr-2.5 text-primary-600 bg-slate-100 border-slate-300 rounded
-                                    focus:ring-primary-500 dark:focus:ring-primary-600 dark:ring-offset-slate-800
-                                    focus:ring-2 dark:bg-slate-700 dark:border-slate-600
-                                    ",
+                                                                                    focus:ring-primary-500 dark:focus:ring-primary-600 dark:ring-offset-slate-800
+                                                                                    focus:ring-2 dark:bg-slate-700 dark:border-slate-600
+                                                                                    ",
                                },
                                true,
                                false %>
@@ -70,14 +76,14 @@
               </div>
               <div>
                 <%= form.hidden_field :export_type, value: "sample" %>
-                <input type="hidden" aria-hidden="true" name="data_export[export_parameters][namespace_id]", value=<%= namespace_id %>>
+                <input type="hidden" aria-hidden="true"
+                name="data_export[export_parameters][namespace_id]", value=<%= namespace_id %>>
               </div>
-              <div class="hidden" data-infinite-scroll-target="all"></div>
               <div>
                 <%= form.submit t(".submit_button"),
                             data: {
                               turbo_frame: "_top",
-                              action: "click->infinite-scroll#clear",
+                              action: "click->form--hidden-inputs#clearSelection",
                             },
                             class: "button button--state-primary button--size-default" %>
               </div>
