+++ app/views/projects/samples/clones/_dialog.html.erb
@@ -5,7 +5,6 @@
       <div
         data-controller="infinite-scroll"
         data-infinite-scroll-selection-outlet='#samples-table'
-        data-infinite-scroll-field-name-value="clone[sample_ids][]"
         data-infinite-scroll-paged-field-name-value="sample_ids[]"
         data-infinite-scroll-singular-value="<%= t(".description.singular") %>"
         data-infinite-scroll-plural-value="<%= t(".description.plural") %>"
@@ -46,7 +45,9 @@
 
           <%= form_for(:clone, url: namespace_project_samples_clone_path, method: :post,
           data: {
-            controller: "spinner",
+            controller: "spinner form--hidden-inputs",
+            "form--hidden-inputs-selection-outlet": '#samples-table',
+            "form--hidden-inputs-field-name-value": "clone[sample_ids][]",
             action:"turbo:submit-start->spinner#submitStart turbo:submit-end->spinner#submitEnd"
           }
         ) do |form| %>
@@ -55,13 +56,12 @@
                 <%= form.label :new_project_id, t(".new_project_id") %>
                 <%= form.collection_select(:new_project_id, @projects, :id, :full_path) %>
               </div>
-              <div class="hidden" data-infinite-scroll-target="all"></div>
               <div>
                 <%= form.submit t(".submit_button"),
                             class: "button button--size-default button--state-primary",
                             disabled: @projects.count.zero?,
                             data: {
-                              action: "click->infinite-scroll#clear",
+                              action: "click->form--hidden-inputs#clearSelection",
                               "spinner-target": "submit",
                             } %>
               </div>
