+++ app/views/projects/samples/clones/create.turbo_stream.erb
@@ -12,3 +12,5 @@
                          errors: errors,
                        } %>
 <% end %>
+
+<turbo-stream action="refresh"></turbo-stream>
