+++ app/views/projects/samples/transfers/_dialog.html.erb
@@ -5,7 +5,6 @@
       <div
         data-controller="infinite-scroll"
         data-infinite-scroll-selection-outlet='#samples-table'
-        data-infinite-scroll-field-name-value="transfer[sample_ids][]"
         data-infinite-scroll-paged-field-name-value="sample_ids[]"
         data-infinite-scroll-singular-value="<%= t(".description.singular") %>"
         data-infinite-scroll-plural-value="<%= t(".description.plural") %>"
@@ -14,6 +13,7 @@
         url: list_namespace_project_samples_path,
         data: { "infinite-scroll-target": "pageForm" }
       ) do %>
+          <input type="hidden" name="format" value="turbo_stream"/>
           <div data-infinite-scroll-target="pageFormContent"></div>
         <% end %>
 
@@ -46,8 +46,10 @@
 
           <%= form_for(:transfer, url: namespace_project_samples_transfer_path, method: :post,
               data: {
-                controller: "spinner",
-                action:"turbo:submit-start->spinner#submitStart turbo:submit-end->spinner#submitEnd"
+                controller: "spinner form--hidden-inputs",
+                "form--hidden-inputs-selection-outlet": '#samples-table',
+                "form--hidden-inputs-field-name-value": "transfer[sample_ids][]",
+                action: "turbo:submit-start->spinner#submitStart turbo:submit-end->spinner#submitEnd"
               }
             ) do |form| %>
             <div class="grid gap-4">
@@ -55,13 +57,12 @@
                 <%= form.label :new_project_id, t(".new_project_id") %>
                 <%= form.collection_select(:new_project_id, @projects, :id, :full_path) %>
               </div>
-              <div class="hidden" data-infinite-scroll-target="all"></div>
               <div>
                 <%= form.submit t(".submit_button"),
                             class: "button button--size-default button--state-primary",
                             disabled: @projects.count.zero?,
                             data: {
-                              action: "click->infinite-scroll#clear",
+                              action: "click->form--hidden-inputs#clearSelection",
                               "spinner-target": "submit",
                             } %>
               </div>
