+++ app/views/projects/samples/transfers/create.turbo_stream.erb
@@ -12,3 +12,5 @@
                          errors: errors,
                        } %>
 <% end %>
+
+<turbo-stream action="refresh"></turbo-stream>
